#143 refactor: context 디렉토리를 관련 파일 기반 폴더 구조로 개선
