+++ src/context/modal/index.ts
@@ -1,0 +1,5 @@
+import { modalContext } from '@/context/modal/context';
+import { ModalProvider } from '@/context/modal/provider';
+import { useModalContext } from '@/context/modal/useModalContext';
+
+export { modalContext, ModalProvider, useModalContext };
